+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ## v0.0.3
+### Added
+* Allow controlling `QualitySettings.asyncUploadPersistentBuffer` via the config.
+
 ### Fixed
 * Fixed an infinite hang when synchronously loading a PNG/JPEG texture.